Actions update

+++ src/wow/action/Action.java
@@ -6,5 +6,5 @@ import wow.player.Player;
 public abstract class Action {
 	public abstract boolean run(Player p, Environment e);
 	public abstract boolean isrunnable(Player p, Environment e);
-	public abstract int potentielUtility(Player p, Environment e);
+	public abstract double potentielUtility(Player p, Environment e);
 }
